Refactoring and log cleaning
