+++ src/graphql/schemaManager.ts
@@ -6,7 +6,7 @@
  */
 
 import type { GraphQLSchema } from "graphql";
-import { pluginLogger } from "~/src/plugin/logger";
+
 import { getPluginManagerInstance } from "~/src/plugin/registry";
 import { builder } from "./builder";
 
@@ -21,40 +21,23 @@ class GraphQLSchemaManager {
 	private async setupPluginListeners(): Promise<void> {
 		const pluginManager = getPluginManagerInstance();
 		if (pluginManager) {
-			await pluginLogger.info("🔗 Setting up Schema Manager Plugin Listeners", {
-				hasPluginManager: !!pluginManager,
-				pluginManagerInitialized: pluginManager.isSystemInitialized(),
-			});
+			console.log("🔗 Setting up Schema Manager Plugin Listeners");
 
 			// Listen for schema rebuild events
 			pluginManager.on("schema:rebuild", async (data) => {
-				await pluginLogger.info("🔄 Schema Rebuild Triggered", {
-					pluginId: data.pluginId,
-					reason: data.reason,
-					timestamp: data.timestamp,
-				});
+				console.log("🔄 Schema Rebuild Triggered", data.pluginId, data.reason);
 				await this.rebuildSchema();
 			});
 
 			// Listen for plugin deactivation to remove fields
 			pluginManager.on("plugin:deactivated", async (pluginId) => {
-				await pluginLogger.info("🔄 Schema Rebuild for Plugin Deactivation", {
-					pluginId,
-					reason: "plugin_deactivation",
-				});
+				console.log("🔄 Schema Rebuild for Plugin Deactivation", pluginId);
 				await this.rebuildSchema();
 			});
 
-			await pluginLogger.info(
-				"✅ Schema Manager Plugin Listeners Setup Complete",
-			);
+			console.log("✅ Schema Manager Plugin Listeners Setup Complete");
 		} else {
-			await pluginLogger.warn(
-				"⚠️ Plugin Manager Not Available for Schema Listeners",
-				{
-					retryStrategy: "will_setup_during_initial_build",
-				},
-			);
+			console.log("⚠️ Plugin Manager Not Available for Schema Listeners");
 		}
 	}
 
@@ -62,9 +45,7 @@ class GraphQLSchemaManager {
 	 * Build the initial schema
 	 */
 	async buildInitialSchema(): Promise<GraphQLSchema> {
-		await pluginLogger.info("🏗️ Building Initial GraphQL Schema", {
-			timestamp: new Date().toISOString(),
-		});
+		console.log("🏗️ Building Initial GraphQL Schema");
 
 		// Set up plugin listeners now that we're initializing
 		await this.setupPluginListeners();
@@ -79,16 +60,7 @@ class GraphQLSchemaManager {
 		const schema = builder.toSchema();
 		this.currentSchema = schema;
 
-		await pluginLogger.info("✅ Initial GraphQL Schema Built", {
-			schemaFields: {
-				queries: Object.keys(schema.getQueryType()?.getFields() || {}),
-				mutations: Object.keys(schema.getMutationType()?.getFields() || {}),
-				subscriptions: Object.keys(
-					schema.getSubscriptionType()?.getFields() || {},
-				),
-			},
-			timestamp: new Date().toISOString(),
-		});
+		console.log("✅ Initial GraphQL Schema Built");
 
 		return schema;
 	}
@@ -98,9 +70,7 @@ class GraphQLSchemaManager {
 	 */
 	async rebuildSchema(): Promise<GraphQLSchema> {
 		if (this.isRebuilding) {
-			await pluginLogger.info("⏳ Schema Rebuild Already In Progress", {
-				action: "skipping_duplicate_rebuild",
-			});
+			console.log("⏳ Schema Rebuild Already In Progress");
 			if (!this.currentSchema) {
 				throw new Error("No current schema available during rebuild");
 			}
@@ -110,10 +80,7 @@ class GraphQLSchemaManager {
 		this.isRebuilding = true;
 
 		try {
-			await pluginLogger.info("🔧 Starting Dynamic Schema Rebuild", {
-				timestamp: new Date().toISOString(),
-				currentSchemaExists: !!this.currentSchema,
-			});
+			console.log("🔧 Starting Dynamic Schema Rebuild");
 
 			// Use the main builder instance (builders maintain global state)
 
@@ -127,35 +94,14 @@ class GraphQLSchemaManager {
 			const newSchema = builder.toSchema();
 			this.currentSchema = newSchema;
 
-			await pluginLogger.info("✅ Schema Rebuild Completed", {
-				newSchemaFields: {
-					queries: Object.keys(newSchema.getQueryType()?.getFields() || {}),
-					mutations: Object.keys(
-						newSchema.getMutationType()?.getFields() || {},
-					),
-					subscriptions: Object.keys(
-						newSchema.getSubscriptionType()?.getFields() || {},
-					),
-				},
-				timestamp: new Date().toISOString(),
-			});
+			console.log("✅ Schema Rebuild Completed");
 
 			// Notify all registered callbacks about the schema update
 			this.notifySchemaUpdateCallbacks(newSchema);
 
 			return newSchema;
 		} catch (error) {
-			await pluginLogger.error("❌ Schema Rebuild Failed", {
-				error:
-					error instanceof Error
-						? {
-								message: error.message,
-								stack: error.stack,
-								name: error.name,
-							}
-						: String(error),
-				timestamp: new Date().toISOString(),
-			});
+			console.error("❌ Schema Rebuild Failed", error);
 			throw error;
 		} finally {
 			this.isRebuilding = false;
@@ -177,13 +123,9 @@ class GraphQLSchemaManager {
 			await import("./types/index");
 			// Note: interfaces and unions directories have empty index files, so skipping them
 
-			await pluginLogger.info("📦 Core Schema Components Imported", {
-				componentsLoaded: ["enums", "scalars", "inputs", "types"],
-			});
+			console.log("📦 Core Schema Components Imported");
 		} catch (error) {
-			await pluginLogger.error("❌ Core Schema Import Failed", {
-				error: error instanceof Error ? error.message : String(error),
-			});
+			console.error("❌ Core Schema Import Failed", error);
 			throw error;
 		}
 	}
@@ -194,17 +136,15 @@ class GraphQLSchemaManager {
 	private async registerActivePluginExtensions(): Promise<void> {
 		const pluginManager = getPluginManagerInstance();
 		if (!pluginManager) {
-			await pluginLogger.warn("Plugin Manager Not Available", {
-				action: "skipping_plugin_extensions",
-			});
+			console.log("Plugin Manager Not Available");
 			return;
 		}
 
 		const activePlugins = pluginManager.getActivePlugins();
-		await pluginLogger.info("🔌 Registering Active Plugin Extensions", {
-			activePluginCount: activePlugins.length,
-			activePluginIds: activePlugins.map((p) => p.id),
-		});
+		console.log(
+			"🔌 Registering Active Plugin Extensions",
+			activePlugins.length,
+		);
 
 		const extensionRegistry = pluginManager.getExtensionRegistry();
 
@@ -268,12 +208,7 @@ class GraphQLSchemaManager {
 			}
 		}
 
-		await pluginLogger.info("✅ Plugin Extensions Registered", {
-			totalQueries: Object.keys(extensionRegistry.graphql.queries).length,
-			totalMutations: Object.keys(extensionRegistry.graphql.mutations).length,
-			totalSubscriptions: Object.keys(extensionRegistry.graphql.subscriptions)
-				.length,
-		});
+		console.log("✅ Plugin Extensions Registered");
 	}
 
 	/**
@@ -296,12 +231,11 @@ class GraphQLSchemaManager {
 					t.string({
 						description: `Plugin ${pluginId} query: ${fieldName}`,
 						resolve: async (parent, args, ctx) => {
-							await pluginLogger.info("⚡ Executing Plugin GraphQL Query", {
+							console.log(
+								"⚡ Executing Plugin GraphQL Query",
 								pluginId,
-								queryName: fieldName,
-								namespacedFieldName,
-								timestamp: new Date().toISOString(),
-							});
+								fieldName,
+							);
 
 							// Create plugin context from the main GraphQL context
 							const pluginContext = {
@@ -337,12 +271,11 @@ class GraphQLSchemaManager {
 							}),
 						},
 						resolve: async (parent, args, ctx) => {
-							await pluginLogger.info("⚡ Executing Plugin GraphQL Mutation", {
+							console.log(
+								"⚡ Executing Plugin GraphQL Mutation",
 								pluginId,
-								mutationName: fieldName,
-								namespacedFieldName,
-								timestamp: new Date().toISOString(),
-							});
+								fieldName,
+							);
 
 							// Parse input if provided and structure it properly for plugin resolvers
 							let formattedArgs = {};
@@ -390,14 +323,10 @@ class GraphQLSchemaManager {
 							return subscriptionGenerator();
 						},
 						resolve: async (parent, args, ctx) => {
-							await pluginLogger.info(
+							console.log(
 								"⚡ Executing Plugin GraphQL Subscription",
-								{
-									pluginId,
-									subscriptionName: fieldName,
-									namespacedFieldName,
-									timestamp: new Date().toISOString(),
-								},
+								pluginId,
+								fieldName,
 							);
 
 							// Create plugin context from the main GraphQL context
@@ -425,13 +354,13 @@ class GraphQLSchemaManager {
 				);
 			}
 		} catch (error) {
-			pluginLogger.error("❌ Failed to Register Plugin GraphQL Field", {
+			console.error(
+				"❌ Failed to Register Plugin GraphQL Field",
 				pluginId,
-				fieldType: type,
+				type,
 				fieldName,
-				namespacedFieldName,
-				error: error instanceof Error ? error.message : String(error),
-			});
+				error,
+			);
 		}
 	}
 
@@ -459,9 +388,7 @@ class GraphQLSchemaManager {
 			try {
 				callback(schema);
 			} catch (error) {
-				pluginLogger.error("❌ Schema Update Callback Failed", {
-					error: error instanceof Error ? error.message : String(error),
-				});
+				console.error("❌ Schema Update Callback Failed", error);
 			}
 		}
 	}
