+++ src/graphql/types/Mutation/plugin/index.ts
@@ -1,0 +1,4 @@
+// Import plugin mutation files to register them with the GraphQL builder
+import "./createPlugin";
+import "./updatePlugin";
+import "./deletePlugin";
