--- test/graphql/types/Mutation/updatePlugin.test.ts
@@ -1,30 +1,0 @@
-import { beforeEach, describe, expect, it, vi } from "vitest";
-import type PluginManager from "../../../../src/plugin/manager";
-import { getPluginManagerInstance } from "../../../../src/plugin/registry";
-
-// Mock the plugin registry
-vi.mock("../../../../src/plugin/registry", () => ({
-	getPluginManagerInstance: vi.fn(),
-}));
-
-describe("updatePlugin mutation", () => {
-	let mockPluginManager: Partial<PluginManager>;
-
-	beforeEach(() => {
-		mockPluginManager = {
-			loadPlugin: vi.fn(),
-			unloadPlugin: vi.fn(),
-			activatePlugin: vi.fn(),
-			deactivatePlugin: vi.fn(),
-		};
-
-		vi.mocked(getPluginManagerInstance).mockReturnValue(
-			mockPluginManager as PluginManager,
-		);
-	});
-
-	it("should be defined in the schema", () => {
-		// Since the resolver is not exported, we just test that the file loads without errors
-		expect(true).toBe(true);
-	});
-});
